+++ tests/real_types_test/src/001_array_key_exists.php
@@ -17,5 +17,6 @@ function test1($x) {
     if (is_array(key($x))) {
         echo "Impossible, key() only returns int|string|null\n";
     }
+    return $x;
 }
 test1([]);